Compare the actual lexer results as a list instead of one by one.
Got rid of the doTest method and created a lex method that lexes the given text and returns a list of the lexed tokens. Then we can compare the list to the expected list. This gives a more useful answer instead of comparing each token one by one and not knowing where it failed.